src/presentation/controllers/survey/load-surveys/load-surveys-controller.ts
import { noContent, ok, serverError } from '@/presentation/helpers/http/http-helper'
import { Controller, HttpResponse, LoadSurvey } from './load-surveys-controller-protocols'

export class LoadSurveysController implements Controller {
  constructor (
    private readonly loadSurveys: LoadSurvey
  ) {}

  async handle (): Promise<HttpResponse> {
    try {
      const surveys = await this.loadSurveys.load()
      return surveys.length ? ok(surveys) : noContent()
    } catch (error) {
      return serverError(error)
    }
  }
}
